# BackEnd/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"WalletDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"currency": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"currency",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateWalletBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"currency": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"currency"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_CreateWalletBody_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"currency": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UpdateWalletBody": {
				"$ref": "#/components/schemas/Partial_CreateWalletBody_"
			},
			"WalletMemberDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WalletInviteDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"walletId": {
						"type": "string"
					},
					"inviterId": {
						"type": "string"
					},
					"inviteeEmail": {
						"type": "string"
					},
					"token": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"expiresAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"walletId",
					"inviterId",
					"inviteeEmail",
					"token",
					"status",
					"expiresAt",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateInviteBody": {
				"properties": {
					"inviteeEmail": {
						"type": "string"
					}
				},
				"required": [
					"inviteeEmail"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"auth0Id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"auth0Id",
					"name",
					"email",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"WalletMemberLinkDto": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"walletId": {
						"type": "string"
					},
					"role": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"walletId",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"sessionCookie": {
				"type": "apiKey",
				"in": "cookie",
				"name": "appSession",
				"description": "Log in at /login in the same browser before trying protected routes."
			}
		}
	},
	"info": {
		"title": "ledger-flow-backend",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
		"/api/wallets": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WalletDto"
								}
							}
						}
					}
				},
				"tags": [
					"Wallets"
				],
				"security": [
					{
						"sessionCookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateWalletBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/WalletDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Wallets"
				],
				"security": [
					{
						"sessionCookie": []
					}
				],
				"parameters": []
			}
		},
		"/api/wallets/{walletId}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WalletDto"
								}
							}
						}
					}
				},
				"tags": [
					"Wallets"
				],
				"security": [
					{
						"sessionCookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "walletId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WalletDto"
								}
							}
						}
					}
				},
				"tags": [
					"Wallets"
				],
				"security": [
					{
						"sessionCookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "walletId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateWalletBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WalletDto"
								}
							}
						}
					}
				},
				"tags": [
					"Wallets"
				],
				"security": [
					{
						"sessionCookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "walletId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/wallets/{walletId}/members": {
			"get": {
				"operationId": "ListMembers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/WalletMemberDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Wallet Members"
				],
				"security": [
					{
						"sessionCookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "walletId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/wallets/{walletId}/invites": {
			"post": {
				"operationId": "CreateWalletInvite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WalletInviteDto"
								}
							}
						}
					}
				},
				"tags": [
					"Wallet Members"
				],
				"security": [
					{
						"sessionCookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "walletId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateInviteBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListWalletInvites",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/WalletInviteDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Wallet Members"
				],
				"security": [
					{
						"sessionCookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "walletId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/wallets/{walletId}/invites/{inviteId}": {
			"delete": {
				"operationId": "RevokeWalletInvite",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Wallet Members"
				],
				"security": [
					{
						"sessionCookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "walletId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "inviteId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/wallets/{walletId}/members/{userId}": {
			"delete": {
				"operationId": "RemoveWalletMember",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Wallet Members"
				],
				"security": [
					{
						"sessionCookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "walletId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/users/me": {
			"post": {
				"operationId": "Sync",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			},
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"sessionCookie": []
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"sessionCookie": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserBody"
							}
						}
					}
				}
			}
		},
		"/api/invites/mine": {
			"get": {
				"operationId": "ListMyInvites",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/WalletInviteDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Invites"
				],
				"security": [
					{
						"sessionCookie": []
					}
				],
				"parameters": []
			}
		},
		"/api/invites/{token}/accept": {
			"post": {
				"operationId": "AcceptWalletInvite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WalletMemberLinkDto"
								}
							}
						}
					}
				},
				"tags": [
					"Invites"
				],
				"security": [
					{
						"sessionCookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/invites/{token}/decline": {
			"post": {
				"operationId": "DeclineWalletInvite",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Invites"
				],
				"security": [
					{
						"sessionCookie": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}